backend/requirements.txt
tqdm
ollama
streamlit
requests
langchain
langchain-ollama
tavily-python
langgraph
pydantic
python-dotenv
clarifai-grpc
fastapi[standard]
sqlalchemy
alembic
uvicorn
passlib[bcrypt]
psycopg2-binary
python-jose[cryptography]
python-multipart
langchain-google-genai
langchain-openai
pytube